Fix Rivermax build errors
The shared manager.cpp and manager.h depend on DPDK headers so the
Rivermax build target requires DPDK to be present.

- Dockerfile - change the rivermax stage to extend from dpdk instead of
  base-deps.
- manager.cpp DPDK include guard only covered DPDK/GPUNetIO/RDMA builds.
  Extend it to cover ANO_MGR_RIVERMAX as well.
- RivermaxMgr::get_segment_packet_length() and get_packet_length() had
  uint16_t return types, mismatching the uint32_t declared in the
  Manager base class.

+++ operators/advanced_network/advanced_network/managers/rivermax/rivermax_mgr_impl/adv_network_rivermax_mgr.cpp
@@ -702,11 +702,11 @@ void* RivermaxMgr::get_packet_ptr(BurstParams* burst, int idx) {
   return pImpl->get_packet_ptr(burst, idx);
 }
 
-uint16_t RivermaxMgr::get_segment_packet_length(BurstParams* burst, int seg, int idx) {
+uint32_t RivermaxMgr::get_segment_packet_length(BurstParams* burst, int seg, int idx) {
   return pImpl->get_segment_packet_length(burst, seg, idx);
 }
 
-uint16_t RivermaxMgr::get_packet_length(BurstParams* burst, int idx) {
+uint32_t RivermaxMgr::get_packet_length(BurstParams* burst, int idx) {
   return pImpl->get_packet_length(burst, idx);
 }
